Consolidate validation/field action service calls into mapping management service. Fail gracefully if mapping service isn't defined.
